fix: can't delete owner from a project

+++ packages/shared/src/queries/projects/delete-member.ts
@@ -1,1 +1,1 @@
-export default `DELETE FROM projects_users_roles WHERE project_id = $1 AND user_id = $2`
+export default `DELETE FROM projects_users_roles WHERE project_id = $1 AND user_id = $2 AND role_id != 1`
